Group clauses in many files
