+++ src/GuardClauses/GuardAgainstNegativeExtensions.cs
@@ -1,0 +1,216 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Diagnostics.CodeAnalysis;
+using System.Linq;
+using System.Text.RegularExpressions;
+using JetBrainsNoEnumerationAttribute = JetBrains.Annotations.NoEnumerationAttribute;
+using JetBrainsNotNullAttribute = JetBrains.Annotations.NotNullAttribute;
+using JetBrainsRegexPattern = JetBrains.Annotations.RegexPatternAttribute;
+using JetBrainsInvokerParameterNameAttribute = JetBrains.Annotations.InvokerParameterNameAttribute;
+
+namespace Ardalis.GuardClauses
+{
+    public static partial class GuardClauseExtensions
+    {
+        /// <summary>
+        /// Throws an <see cref="ArgumentException" /> if <paramref name="input"/> is negative.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not negative.</returns>
+        /// <exception cref="ArgumentException"></exception>
+        public static int Negative([JetBrainsNotNull] this IGuardClause guardClause, int input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return Negative<int>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException" /> if <paramref name="input"/> is negative.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not negative.</returns>
+        /// <exception cref="ArgumentException"></exception>
+        public static long Negative([JetBrainsNotNull] this IGuardClause guardClause, long input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return Negative<long>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException" /> if <paramref name="input"/> is negative.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not negative.</returns>
+        /// <exception cref="ArgumentException"></exception>
+        public static decimal Negative([JetBrainsNotNull] this IGuardClause guardClause, decimal input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return Negative<decimal>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException" /> if <paramref name="input"/> is negative.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not negative.</returns>
+        /// <exception cref="ArgumentException"></exception>
+        public static float Negative([JetBrainsNotNull] this IGuardClause guardClause, float input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return Negative<float>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException" /> if <paramref name="input"/> is negative.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not negative.</returns>
+        /// <exception cref="ArgumentException"></exception>
+        public static double Negative([JetBrainsNotNull] this IGuardClause guardClause, double input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return Negative<double>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException" /> if <paramref name="input"/> is negative.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not negative.</returns>
+        /// <exception cref="ArgumentException"></exception>
+        public static TimeSpan Negative([JetBrainsNotNull] this IGuardClause guardClause, TimeSpan input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return Negative<TimeSpan>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException" /> if <paramref name="input"/> is negative.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not negative.</returns>
+        /// <exception cref="ArgumentException"></exception>
+        private static T Negative<T>([JetBrainsNotNull] this IGuardClause guardClause, T input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null) where T : struct, IComparable
+        {
+            if (input.CompareTo(default(T)) < 0)
+            {
+                throw new ArgumentException(message ?? $"Required input {parameterName} cannot be negative.", parameterName);
+            }
+
+            return input;
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException"/> if <paramref name="input"/> is negative or zero.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not negative or zero.</returns>
+        public static int NegativeOrZero([JetBrainsNotNull] this IGuardClause guardClause, int input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return NegativeOrZero<int>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException"/> if <paramref name="input"/> is negative or zero.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not negative or zero.</returns>
+        public static long NegativeOrZero([JetBrainsNotNull] this IGuardClause guardClause, long input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return NegativeOrZero<long>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException"/> if <paramref name="input"/> is negative or zero.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not negative or zero.</returns>
+        public static decimal NegativeOrZero([JetBrainsNotNull] this IGuardClause guardClause, decimal input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return NegativeOrZero<decimal>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException"/> if <paramref name="input"/> is negative or zero.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not negative or zero.</returns>
+        public static float NegativeOrZero([JetBrainsNotNull] this IGuardClause guardClause, float input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return NegativeOrZero<float>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException"/> if <paramref name="input"/> is negative or zero.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not negative or zero.</returns>
+        public static double NegativeOrZero([JetBrainsNotNull] this IGuardClause guardClause, double input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return NegativeOrZero<double>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException"/> if <paramref name="input"/> is negative or zero.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not negative or zero.</returns>
+        public static TimeSpan NegativeOrZero([JetBrainsNotNull] this IGuardClause guardClause, TimeSpan input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return NegativeOrZero<TimeSpan>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException"/> if <paramref name="input"/> is negative or zero. 
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not negative or zero.</returns>
+        private static T NegativeOrZero<T>([JetBrainsNotNull] this IGuardClause guardClause, T input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null) where T : struct, IComparable
+        {
+            if (input.CompareTo(default(T)) <= 0)
+            {
+                throw new ArgumentException(message ?? $"Required input {parameterName} cannot be zero or negative.", parameterName);
+            }
+
+            return input;
+        }
+    }
+}
