+++ src/GuardClauses/GuardAgainstZeroExtensions.cs
@@ -1,0 +1,118 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Diagnostics.CodeAnalysis;
+using System.Linq;
+using System.Text.RegularExpressions;
+using JetBrainsNoEnumerationAttribute = JetBrains.Annotations.NoEnumerationAttribute;
+using JetBrainsNotNullAttribute = JetBrains.Annotations.NotNullAttribute;
+using JetBrainsRegexPattern = JetBrains.Annotations.RegexPatternAttribute;
+using JetBrainsInvokerParameterNameAttribute = JetBrains.Annotations.InvokerParameterNameAttribute;
+
+namespace Ardalis.GuardClauses
+{
+    public static partial class GuardClauseExtensions
+    {
+        /// <summary>
+        /// Throws an <see cref="ArgumentException" /> if <paramref name="input" /> is zero.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not zero.</returns>
+        /// <exception cref="ArgumentException"></exception>
+        public static int Zero([JetBrainsNotNull] this IGuardClause guardClause, int input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return Zero<int>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException" /> if <paramref name="input" /> is zero.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not zero.</returns>
+        /// <exception cref="ArgumentException"></exception>
+        public static long Zero([JetBrainsNotNull] this IGuardClause guardClause, long input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return Zero<long>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException" /> if <paramref name="input" /> is zero.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not zero.</returns>
+        /// <exception cref="ArgumentException"></exception>
+        public static decimal Zero([JetBrainsNotNull] this IGuardClause guardClause, decimal input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return Zero<decimal>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException" /> if <paramref name="input" /> is zero.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not zero.</returns>
+        /// <exception cref="ArgumentException"></exception>
+        public static float Zero([JetBrainsNotNull] this IGuardClause guardClause, float input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return Zero<float>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException" /> if <paramref name="input" /> is zero.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not zero.</returns>
+        /// <exception cref="ArgumentException"></exception>
+        public static double Zero([JetBrainsNotNull] this IGuardClause guardClause, double input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null)
+        {
+            return Zero<double>(guardClause, input, parameterName, message);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException" /> if <paramref name="input" /> is zero.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <returns><paramref name="input" /> if the value is not zero.</returns>
+        /// <exception cref="ArgumentException"></exception>
+        public static TimeSpan Zero([JetBrainsNotNull] this IGuardClause guardClause, TimeSpan input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName)
+        {
+            return Zero<TimeSpan>(guardClause, input, parameterName);
+        }
+
+        /// <summary>
+        /// Throws an <see cref="ArgumentException" /> if <paramref name="input" /> is zero.
+        /// </summary>
+        /// <param name="guardClause"></param>
+        /// <param name="input"></param>
+        /// <param name="parameterName"></param>
+        /// <param name="message">Optional. Custom error message</param>
+        /// <returns><paramref name="input" /> if the value is not zero.</returns>
+        /// <exception cref="ArgumentException"></exception>
+        private static T Zero<T>([JetBrainsNotNull] this IGuardClause guardClause, T input, [JetBrainsNotNull][JetBrainsInvokerParameterName] string parameterName, string? message = null) where T : struct
+        {
+            if (EqualityComparer<T>.Default.Equals(input, default(T)))
+            {
+                throw new ArgumentException(message ?? $"Required input {parameterName} cannot be zero.", parameterName);
+            }
+
+            return input;
+        }
+    }
+}
